Promote mixed signed/unsigned integer operands safely
Binary arithmetic and comparison operators now use the same lossless
rule as assignment for mixed signed/unsigned integer operands: both
operands promote to the narrowest signed type that holds their full
ranges. A strictly-wider signed type covers an unsigned operand
(ubyte + int is int, uint + long is long), and equal-width or narrower
signed operands widen to the next signed width (int + uint is long,
byte + ubyte is short). Mixing with a 64-bit unsigned type such as
ulong or ptr_uint has no safe signed common type and still requires an
explicit cast. The lowering's common-integer-type copy mirrors the
analyzer rule so the emitted C casts match the promoted result type.
Bitwise and shift operators stay strict.

+++ test/compiler/semantic/error_detection_test.rb
@@ -829,7 +829,7 @@ class ErrorDetectionTest < Minitest::Test
 
       function main() -> int:
           let left: int = 1
-          let right: uint = 2
+          let right: ulong = 2
           let sum = left + right
           return sum
     MT
